refactor(website): move CLI client ids into config/cliClients
