+++ apps/website/src/scripts/posthog.ts
@@ -3,6 +3,7 @@ import posthog from 'posthog-js'
 import { createPostHogBeforeSend } from '@comfyorg/shared-frontend-utils/piiUtil'
 
 import type { Platform } from '@/composables/useDownloadUrl'
+import type { CliClientId } from '@/config/cliClients'
 import type { ConnectionId, McpClientId } from '@/config/mcpClients'
 
 const POSTHOG_KEY =
@@ -21,29 +22,6 @@ const ANALYTICS_EVENT = {
   mcpConnectionTabClicked: 'website:mcp_connection_tab_clicked',
   mcpClientTabClicked: 'website:mcp_client_tab_clicked'
 } as const
-
-const CLI_CLIENT_IDS = [
-  'claude-code',
-  'codex',
-  'cursor',
-  'gemini-cli',
-  'openclaw',
-  'hermes',
-  'terminal',
-  'ci'
-] as const
-
-export type CliClientId = (typeof CLI_CLIENT_IDS)[number]
-
-// Mirrors isMcpClientId (@/config/mcpClients): CliClientId is a static
-// union rather than data-derived, so the guard checks against the id list
-// directly instead of taking a connections record.
-export function isCliClientId(value: unknown): value is CliClientId {
-  return (
-    typeof value === 'string' &&
-    (CLI_CLIENT_IDS as readonly string[]).includes(value)
-  )
-}
 
 type AnalyticsEvent =
   | { name: typeof ANALYTICS_EVENT.pageview; properties?: undefined }
