+++ apps/website/src/templates/cli/SetupSection.vue
@@ -5,14 +5,14 @@ import { ref } from 'vue'
 import SectionHeader from '../../components/common/SectionHeader.vue'
 import SurfaceToggle from '../../components/common/SurfaceToggle.vue'
 import CopyableField from '../../components/ui/copyable-field/CopyableField.vue'
+import type { CliClientId } from '../../config/cliClients'
+import { isCliClientId } from '../../config/cliClients'
 import { externalLinks, getRoutes } from '../../config/routes'
 import type { Locale } from '../../i18n/translations'
 import { t } from '../../i18n/translations'
-import type { CliClientId } from '../../scripts/posthog'
 import {
   captureCliClientTabClick,
-  captureCliConnectionTabClick,
-  isCliClientId
+  captureCliConnectionTabClick
 } from '../../scripts/posthog'
 
 const { locale = 'en' } = defineProps<{ locale?: Locale }>()
